# Wedding/Wedding.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Wedding.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Guest": {
      "get": {
        "tags": [
          "Guest"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filterOn",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filterQuery",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "isAscending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Guest"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateGuestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Guest/{id}": {
      "get": {
        "tags": [
          "Guest"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Guest"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateGuestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Guest"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/GuestList": {
      "get": {
        "tags": [
          "GuestList"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filterOn",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filterQuery",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "isAscending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "post": {
        "tags": [
          "GuestList"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateGuestListDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/GuestList/{id}": {
      "get": {
        "tags": [
          "GuestList"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "GuestList"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateGuestListDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "GuestList"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Invitation": {
      "get": {
        "tags": [
          "Invitation"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filterOn",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filterQuery",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "isAscending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Invitation"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateInvitationDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Invitation/{id}": {
      "get": {
        "tags": [
          "Invitation"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Invitation"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateInvitationDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Invitation"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateGuestDTO": {
      "type": "object",
      "properties": {
        "guestId": {
          "format": "uuid",
          "type": "string"
        },
        "eventId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "attend": {
          "type": "string"
        },
        "gift": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateGuestListDTO": {
      "type": "object",
      "properties": {
        "guestId": {
          "format": "uuid",
          "type": "string"
        },
        "eventId": {
          "format": "uuid",
          "type": "string"
        },
        "guestName": {
          "type": "string"
        },
        "attendStatus": {
          "type": "string"
        },
        "checkinTime": {
          "format": "date-time",
          "type": "string"
        },
        "guestGift": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateInvitationDTO": {
      "type": "object",
      "properties": {
        "invitationId": {
          "format": "uuid",
          "type": "string"
        },
        "weddingId": {
          "format": "uuid",
          "type": "string"
        },
        "templateId": {
          "format": "uuid",
          "type": "string"
        },
        "customerMessage": {
          "type": "string"
        },
        "customerTextColor": {
          "type": "string"
        },
        "shareableLink": {
          "type": "string"
        },
        "createdTime": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "ResponseDTO": {
      "type": "object",
      "properties": {
        "result": { },
        "isSuccess": {
          "type": "boolean"
        },
        "statusCode": {
          "format": "int32",
          "type": "integer"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateGuestDTO": {
      "type": "object",
      "properties": {
        "guestId": {
          "format": "uuid",
          "type": "string"
        },
        "eventId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "attend": {
          "type": "string"
        },
        "gift": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UpdateGuestListDTO": {
      "type": "object",
      "properties": {
        "guestId": {
          "format": "uuid",
          "type": "string"
        },
        "eventId": {
          "format": "uuid",
          "type": "string"
        },
        "guestName": {
          "type": "string"
        },
        "attendStatus": {
          "type": "string"
        },
        "checkinTime": {
          "format": "date-time",
          "type": "string"
        },
        "guestGift": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UpdateInvitationDTO": {
      "type": "object",
      "properties": {
        "invitationId": {
          "format": "uuid",
          "type": "string"
        },
        "weddingId": {
          "format": "uuid",
          "type": "string"
        },
        "templateId": {
          "type": "string"
        },
        "customerMessage": {
          "type": "string"
        },
        "customerTextColor": {
          "type": "string"
        },
        "shareableLink": {
          "type": "string"
        },
        "createdTime": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please enter your token with this format: \"Bearer YOUR_TOKEN\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}